simplify rag factory
