+++ metagpt/rag/rankers/__init__.py
@@ -1,6 +1,1 @@
 """Rankers init"""
-
-from metagpt.rag.rankers.factory import get_rankers
-
-
-__all__ = ["get_rankers"]
