+++ metagpt/rag/retrievers/__init__.py
@@ -1,6 +1,5 @@
 """Retrievers init"""
 
 from metagpt.rag.retrievers.hybrid_retriever import SimpleHybridRetriever
-from metagpt.rag.retrievers.factory import get_retriever
 
-__all__ = ["SimpleHybridRetriever", "get_retriever"]
+__all__ = ["SimpleHybridRetriever"]
